requirements.txt
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
PyYAML==6.0.2
six==1.17.0
